requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pillow==10.0.1
python-dotenv==1.0.0